Project/requirements.txt
langchain>=0.2.0
langgraph==0.2.0
google-generativeai==0.3.2
openai>=1.0.0
python-dotenv==1.0.0
pydantic==2.5.2
streamlit>=1.28.0
typing-extensions>=4.5.0